Changed int to size_t to fix a warning
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/trunk@4261 5e2c2758-810a-0410-8eb3-8e4e92d4004e

+++ src/MPProblem/MPProblem.h
@@ -136,7 +136,7 @@ class MPProblem
     //RemoveObstacleAt
     //a wrapper call to remove a multibody from the environment
     void RemoveObstacleAt(size_t _index);
-    int getSolverNum() { return m_solver.size(); }
+    size_t getSolverNum() { return m_solver.size(); }
 
   protected:
     virtual void Initialize();
